+++ src/framework/piral-native/src/index.ts
@@ -1,6 +1,0 @@
-export * from 'piral-core';
-
-export * from './api';
-export * from './app';
-export * from './render';
-export * from './types';