+++ src/tooling/piral-cli/src/types/public.ts
@@ -235,7 +235,7 @@ export type PackageType = 'registry' | 'file' | 'git';
 
 export type NpmClientType = 'npm' | 'yarn' | 'pnp' | 'pnpm' | 'lerna' | 'rush';
 
-export type Framework = 'piral-native' | 'piral' | 'piral-core' | 'piral-base';
+export type Framework = 'piral' | 'piral-core' | 'piral-base';
 
 export interface StandardEnvProps {
   production?: boolean;