+++ src/utilities/piral-ssr-utils/runtime.js
@@ -1,2 +1,0 @@
-const config = require('./lib/config');
-Object.keys(config).forEach(key => (exports[key] = config[key]));